LPS-64225 Move TreePathUtil to its own package, and deprecated the old one.
